--- a/sources/apps/about/main.c
+++ b/sources/apps/about/main.c
@@ -25,7 +25,7 @@ UiView *about_informations(void)
 
 UiWin *about_win_create(UiApp *app)
 {
-    UiWin *self = ui_win_create(app, m_rect(150, 150, 500, 300), UI_WIN_NORMAL);
+    UiWin *self = ui_win_create(app, m_rectf(150, 150, 500, 300), UI_WIN_NORMAL);
 
     UiView *container = ui_panel_create(UI_COLOR_BASE00);
     ui_view_layout(container, "flex");