--- a/sources/libs/brutal/gfx.h
+++ b/sources/libs/brutal/gfx.h
@@ -7,5 +7,6 @@
 #include <brutal/gfx/dirty.h>
 #include <brutal/gfx/fmt.h>
 #include <brutal/gfx/gfx.h>
+#include <brutal/gfx/ops.h>
 #include <brutal/gfx/paint.h>
 #include <brutal/gfx/path.h>