--- a/sources/libs/brutal/gfx/dirty.c
+++ b/sources/libs/brutal/gfx/dirty.c
@@ -10,18 +10,18 @@ void gfx_dirty_deinit(GfxDirty *self)
     vec_deinit(self);
 }
 
-void gfx_dirty_rect(GfxDirty *self, MRect rect)
+void gfx_dirty_rect(GfxDirty *self, MRectf rect)
 {
     vec_foreach(dirty, self)
     {
-        if (m_rect_collide_rect(*dirty, rect))
+        if (m_rectf_collide_rect(*dirty, rect))
         {
-            MRect rects[4];
-            m_rect_substract(rect, *dirty, rects);
+            MRectf rects[4];
+            m_rectf_substract(rect, *dirty, rects);
 
             for (int i = 0; i < 4; i++)
             {
-                if (!m_rect_empty(rects[i]))
+                if (!m_rectf_empty(rects[i]))
                     gfx_dirty_rect(self, rects[i]);
             }