--- a/sources/libs/brutal/gfx/font.c
+++ b/sources/libs/brutal/gfx/font.c
@@ -27,24 +27,24 @@ GfxFontMeasure gfx_font_measure(GfxFont font, GfxFontStyle style, Str str)
     GfxFontMetrics metrics = gfx_font_metrics(font, style);
 
     return (GfxFontMeasure){
-        .capbound = m_rect(0, 0, len, metrics.ascend + metrics.descend),
-        .linebound = m_rect(0, 0, len, metrics.line_ascend + metrics.line_descend),
-        .baseline = m_vec2(0, metrics.ascend),
+        .capbound = m_rectf(0, 0, len, metrics.ascend + metrics.descend),
+        .linebound = m_rectf(0, 0, len, metrics.line_ascend + metrics.line_descend),
+        .baseline = m_vec2f(0, metrics.ascend),
     };
 }
 
-void gfx_font_render_rune(GfxFont font, GfxFontStyle style, struct _Gfx *gfx, MVec2 baseline, Rune rune)
+void gfx_font_render_rune(GfxFont font, GfxFontStyle style, struct _Gfx *gfx, MVec2f baseline, Rune rune)
 {
     font.render(font.ctx, style, gfx, baseline, rune);
 }
 
-void gfx_font_render_str(GfxFont font, GfxFontStyle style, struct _Gfx *gfx, MVec2 baseline, Str str)
+void gfx_font_render_str(GfxFont font, GfxFontStyle style, struct _Gfx *gfx, MVec2f baseline, Str str)
 {
     Rune rune;
     while (uft8_next_rune(&str, &rune))
     {
         gfx_font_render_rune(font, style, gfx, baseline, rune);
-        baseline = m_vec2_add(baseline, m_vec2(gfx_font_advance(font, style, rune), 0));
+        baseline = m_vec2f_add(baseline, m_vec2f(gfx_font_advance(font, style, rune), 0));
     }
 }
 
@@ -68,9 +68,9 @@ float gfx_font_builtin_advance(MAYBE_UNUSED void *ctx, MAYBE_UNUSED GfxFontStyle
     return GFX_FONT_BUILTIN_WIDTH * style.scale;
 }
 
-void gfx_font_builtin_render(MAYBE_UNUSED void *ctx, GfxFontStyle style, Gfx *gfx, MVec2 baseline, Rune rune)
+void gfx_font_builtin_render(MAYBE_UNUSED void *ctx, GfxFontStyle style, Gfx *gfx, MVec2f baseline, Rune rune)
 {
-    baseline = m_vec2_add(gfx_peek(gfx)->origin, baseline);
+    baseline = m_vec2f_add(gfx_peek(gfx)->origin, baseline);
 
     GfxFontMetrics metrics = gfx_font_builtin_metrics(ctx, style);
     uint8_t index = cp437_from_rune(rune);
@@ -86,9 +86,9 @@ void gfx_font_builtin_render(MAYBE_UNUSED void *ctx, GfxFontStyle style, Gfx *gf
             uint8_t byte = gfx_font_builtin_data[yy];
             uint8_t bit = (byte >> xx) & 0b1;
 
-            MVec2 pos = m_vec2_add(baseline, m_vec2(x, y - (metrics.ascend)));
+            MVec2f pos = m_vec2f_add(baseline, m_vec2f(x, y - (metrics.ascend)));
 
-            if (bit && m_rect_collide_point(gfx_peek(gfx)->clip, pos))
+            if (bit && m_rectf_collide_point(gfx_peek(gfx)->clip, pos))
             {
                 gfx_buf_blend(gfx->buf, pos.x, pos.y, color);
             }