--- a/sources/libs/brutal/gfx/ops.c
+++ b/sources/libs/brutal/gfx/ops.c
@@ -89,7 +89,7 @@ static inline void gfx_ops_clear_common(GfxBuf self, GfxColor color, int const s
     }
 }
 
-void gfx_ops_clear(GfxBuf self, GfxColor color)
+FLATTEN void gfx_ops_clear(GfxBuf self, GfxColor color)
 {
     size_t size = gfx_fmt_size(self.fmt);
 
@@ -136,7 +136,7 @@ static inline void gfx_ops_rect_common(GfxBuf self, MRectf rect, GfxColor color,
     }
 }
 
-void gfx_ops_rect(GfxBuf self, MRectf rect, GfxColor color)
+FLATTEN void gfx_ops_rect(GfxBuf self, MRectf rect, GfxColor color)
 {
     size_t size = gfx_fmt_size(self.fmt);