… sources/apps/bench/bench.h
… sources/apps/bench/benchs/rect.c
… sources/apps/bench/benchs/text.c
… sources/apps/bench/main.c
… sources/libs/brutal/gfx/gfx.h
… sources/libs/brutal/gfx/paint.h
… sources/libs/brutal/gfx/path.c
… sources/libs/brutal/gfx/path.h
… sources/libs/brutal/gfx/stroke.c
… sources/libs/brutal/math/dock.h
… sources/libs/brutal/math/edge.h
… sources/libs/brutal/math/flow.c
… sources/libs/brutal/math/flow.h
… sources/libs/brutal/math/gravity.c
… sources/libs/brutal/math/gravity.h
… sources/libs/brutal/math/rect.h
… sources/libs/brutal/math/spacing.c
… sources/libs/brutal/math/spacing.h
… sources/libs/brutal/math/trans2.h
… sources/libs/brutal/math/vec2.h
… sources/libs/brutal/ui/app.c
… sources/libs/brutal/ui/app.h
… sources/libs/brutal/ui/event.h
… sources/libs/brutal/ui/icon.c
… sources/libs/brutal/ui/image.c
… sources/libs/brutal/ui/layout.c
… sources/libs/brutal/ui/layout.h
… sources/libs/brutal/ui/parser.c
… sources/libs/brutal/ui/slider.c
… sources/libs/brutal/ui/text.c