The demo application on my computer: (<!> **the benchmark is not very scientific** <!>)
| /     | before | after |
| ---- | ----- | ---- |
| debug | 30 fps | 175 fps |
| fast | 81 fps | 1351 fps |